+++ src/irmin-unix/resolver.ml
@@ -550,29 +550,35 @@ let load_config ?root ?config_path ?store ?hash ?contents () =
 
 let string_value = function `String s -> s | _ -> raise Not_found
 
+let find_key config name =
+  Yaml.Util.find_exn name config |> Option.map (fun x -> string_value x)
+
+let handle_decode_err err t x =
+  match Irmin.Type.of_string t x with Ok h -> h | _ -> invalid_arg err
+
 let get_branch (type a)
     (module S : Irmin.Generic_key.S with type Schema.Branch.t = a) config branch
     =
-  let of_string = function
-    | Some x -> (
-        match Irmin.Type.of_string S.Branch.t x with
-        | Ok x -> Some x
-        | _ -> None)
-    | None -> None
-  in
+  let of_string = Option.map (handle_decode_err "invalid branch" S.Branch.t) in
   match branch with
-  | None ->
-      let br =
-        Yaml.Util.find_exn "branch" config
-        |> Option.map (fun x -> string_value x)
-      in
-      of_string br
+  | None -> of_string (find_key config "branch")
   | Some t -> of_string (Some t)
 
-let build_irmin_config config root opts (store, hash, contents) branch : store =
+let get_commit (type a b)
+    (module S : Irmin.Generic_key.S
+      with type commit = a
+       and type Schema.Hash.t = b) config commit =
+  let of_string = Option.map (handle_decode_err "invalid commit" S.Hash.t) in
+  match commit with
+  | None -> of_string (find_key config "commit")
+  | Some t -> of_string (Some t)
+
+let build_irmin_config config root opts (store, hash, contents) branch commit :
+    store =
   let (T { impl; spec; remote }) = get_store config (store, hash, contents) in
   let (module S) = Store.Impl.generic_keyed impl in
   let branch = get_branch (module S) config branch in
+  let commit = get_commit (module S) config commit in
   let config = parse_config ?root config spec in
   let config =
     List.fold_left
@@ -594,9 +600,15 @@ let build_irmin_config config root opts (store, hash, contents) branch : store =
       config (List.flatten opts)
   in
   let spec =
-    match branch with
-    | None -> S.Repo.v config >>= S.main
-    | Some b -> S.Repo.v config >>= fun repo -> S.of_branch repo b
+    match (branch, commit) with
+    | _, Some hash -> (
+        S.Repo.v config >>= fun repo ->
+        let* commit = S.Commit.of_hash repo hash in
+        match commit with
+        | None -> invalid_arg "unknown commit"
+        | Some c -> S.of_commit c)
+    | None, None -> S.Repo.v config >>= S.main
+    | Some b, None -> S.Repo.v config >>= fun repo -> S.of_branch repo b
   in
   S (impl, spec, remote)
 
@@ -607,12 +619,20 @@ let branch =
   in
   Arg.(value & opt (some string) None & doc)
 
-let store () =
-  let create store (root, config_path, opts) branch =
-    let y = read_config_file config_path in
-    build_irmin_config y root opts store branch
+let commit =
+  let doc =
+    Arg.info
+      ~doc:"The store's head commit. This will take precedence over --branch."
+      ~docs:global_option_section ~docv:"COMMIT" [ "commit" ]
   in
-  Term.(const create $ Store.term () $ config_term $ branch)
+  Arg.(value & opt (some string) None & doc)
+
+let store () =
+  let create store (root, config_path, opts) branch commit =
+    let y = read_config_file config_path in
+    build_irmin_config y root opts store branch commit
+  in
+  Term.(const create $ Store.term () $ config_term $ branch $ commit)
 
 let header_conv =
   let parse str =